+++ tests/Tree.spec.tsx
@@ -1,0 +1,1123 @@
+/* eslint-disable no-undef, react/no-multi-comp, no-console,
+react/no-unused-state, react/prop-types, no-return-assign */
+import React from 'react';
+import { render, fireEvent } from '@testing-library/react';
+import { resetWarned } from 'rc-util/lib/warning';
+import { spyElementPrototypes } from 'rc-util/lib/test/domHook';
+import VirtualList from 'rc-virtual-list';
+import Tree, { TreeNode } from '../src';
+import { InternalTreeNode } from '../src/TreeNode';
+import { objectMatcher, spyConsole, spyError } from './util';
+import { convertNodePropsToEventData } from '../src/utils/treeUtil';
+
+const OPEN_CLASSNAME = '.rc-tree-switcher_open';
+const CHECKED_CLASSNAME = '.rc-tree-checkbox-checked';
+const SELECTED_CLASSNAME = '.rc-tree-node-selected';
+
+const delay = timeout =>
+  new Promise(resolve => {
+    setTimeout(resolve, timeout);
+  });
+
+describe('Tree Basic', () => {
+  spyConsole();
+
+  it('TreeNode is in Tree', () => {
+    expect(TreeNode).toBe(Tree.TreeNode);
+  });
+
+  it('renders correctly', () => {
+    const { asFragment } = render(
+      <Tree
+        className="forTest"
+        selectable
+        checkable
+        defaultExpandAll
+        showIcon
+        showLine
+        multiple
+        focusable
+      >
+        <TreeNode title="parent 1" key="0-0" className="spe">
+          <TreeNode title="leaf 1" key="0-0-0" disabled>
+            <TreeNode title="leaf" key="random" />
+            <TreeNode title="leaf" />
+            {null /* Supports conditional rendering */}
+          </TreeNode>
+          <TreeNode title="leaf 2" key="0-0-1" disableCheckbox />
+        </TreeNode>
+      </Tree>,
+    );
+    expect(asFragment().firstChild).toMatchSnapshot();
+  });
+
+  // it('switcherIcon = null, no render rc-tree-switcher null', () => {
+  //   const wrapper = mount(
+  //     <Tree className="forTest" defaultExpandAll switcherIcon={() => false}>
+  //       <TreeNode title="parent 1" key="0-0" className="spe">
+  //         <TreeNode title="leaf 1" key="0-0-0" disabled>
+  //           <TreeNode title="leaf" key="random" />
+  //           <TreeNode title="leaf" />
+  //         </TreeNode>
+  //         <TreeNode title="leaf 2" key="0-0-1" disableCheckbox />
+  //       </TreeNode>
+  //     </Tree>,
+  //   );
+  //   wrapper.update();
+  //   expect(wrapper.find('.rc-tree-switcher').exists()).toBeFalsy();
+  // });
+
+  // describe('expanded', () => {
+  //   it('expands all keys', () => {
+  //     const wrapper = mount(
+  //       <Tree defaultExpandAll>
+  //         <TreeNode title="parent 1" key="0-0">
+  //           <TreeNode title="leaf 1" key="0-0-0" />
+  //         </TreeNode>
+  //       </Tree>,
+  //     );
+  //     const switcher = wrapper.find('.rc-tree-switcher').first();
+  //     expect(switcher.is(OPEN_CLASSNAME)).toBe(true);
+  //   });
+
+  //   it('expands default expanded keys', () => {
+  //     const wrapper = mount(
+  //       <Tree defaultExpandedKeys={['0-0']}>
+  //         <TreeNode title="parent 1" key="0-0">
+  //           <TreeNode title="leaf 1" key="0-0-0" />
+  //         </TreeNode>
+  //       </Tree>,
+  //     );
+  //     const switcher = wrapper.find('.rc-tree-switcher').first();
+  //     expect(switcher.is(OPEN_CLASSNAME)).toBe(true);
+  //   });
+
+  //   it('controlled by expanded keys', () => {
+  //     const wrapper = mount(
+  //       <Tree expandedKeys={[]}>
+  //         <TreeNode title="parent 1" key="0-0">
+  //           <TreeNode title="leaf 1" key="0-0-0" />
+  //         </TreeNode>
+  //       </Tree>,
+  //     );
+  //     const getSwitcher = () => wrapper.find('.rc-tree-switcher').first();
+  //     expect(getSwitcher().is(OPEN_CLASSNAME)).toBe(false);
+  //     wrapper.setProps({ expandedKeys: ['0-0'] });
+  //     expect(getSwitcher().is(OPEN_CLASSNAME)).toBe(true);
+  //   });
+
+  //   it('expands parent node when child node is expanded', () => {
+  //     const wrapper = mount(
+  //       <Tree expandedKeys={['0-0-0']}>
+  //         <TreeNode title="parent 1" key="0-0">
+  //           <TreeNode title="leaf 1" key="0-0-0">
+  //             <TreeNode title="leaf" key="0-0-0-0" />
+  //           </TreeNode>
+  //         </TreeNode>
+  //       </Tree>,
+  //     );
+
+  //     const switcher = wrapper.find('.rc-tree-switcher').first();
+  //     expect(switcher.is(OPEN_CLASSNAME)).toBe(true);
+  //   });
+
+  //   it('use treeData to expand the parent node when the parent node key type is numeric', () => {
+  //     const Demo = () => (
+  //       <Tree
+  //         defaultExpandParent
+  //         defaultExpandedKeys={[22]}
+  //         treeData={[{ key: 11, title: 11, children: [{ key: 22, title: 22 }] }]}
+  //       />
+  //     );
+
+  //     const wrapper = mount(<Demo />);
+
+  //     const switcher = wrapper.find('.rc-tree-switcher').first();
+  //     expect(switcher.is(OPEN_CLASSNAME)).toBe(true);
+  //   });
+
+  //   it('does not expand parent node when autoExpandParent is false', () => {
+  //     const wrapper = mount(
+  //       <Tree expandedKeys={['0-0-0']} defaultExpandParent={false}>
+  //         <TreeNode title="parent 1" key="0-0">
+  //           <TreeNode title="leaf 1" key="0-0-0">
+  //             <TreeNode title="leaf" key="0-0-0-0" />
+  //           </TreeNode>
+  //         </TreeNode>
+  //       </Tree>,
+  //     );
+  //     const switcher = wrapper.find('.rc-tree-switcher').first();
+  //     expect(switcher.is(OPEN_CLASSNAME)).toBe(false);
+  //   });
+
+  //   it('update to expand parent node with autoExpandParent', () => {
+  //     const wrapper = mount(
+  //       <Tree expandedKeys={['0-0-0']} defaultExpandParent={false}>
+  //         <TreeNode title="parent 1" checkable={false} key="0-0">
+  //           <TreeNode title="leaf 1" key="0-0-0">
+  //             <TreeNode title="leaf" key="0-0-0-0" />
+  //           </TreeNode>
+  //         </TreeNode>
+  //       </Tree>,
+  //     );
+  //     let parentSwitcher = wrapper.find('.rc-tree-switcher').first();
+  //     expect(parentSwitcher.is(OPEN_CLASSNAME)).toBe(false);
+
+  //     wrapper.setProps({ autoExpandParent: true });
+
+  //     parentSwitcher = wrapper.find('.rc-tree-switcher').first();
+  //     const childSwitcher = wrapper.find('.rc-tree-switcher').at(1);
+  //     expect(parentSwitcher.is(OPEN_CLASSNAME)).toBe(true);
+  //     expect(childSwitcher.is(OPEN_CLASSNAME)).toBe(true);
+  //   });
+
+  //   it('skip only if disabled with autoExpandParent', () => {
+  //     const wrapper = mount(
+  //       <Tree expandedKeys={['0-0-0-0']} defaultExpandParent>
+  //         <TreeNode title="parent 1" key="0-0">
+  //           <TreeNode title="leaf 1" disabled key="0-0-0">
+  //             <TreeNode title="leaf" key="0-0-0-0" />
+  //           </TreeNode>
+  //         </TreeNode>
+  //       </Tree>,
+  //     );
+
+  //     expect(wrapper.state().expandedKeys.sort()).toEqual(['0-0-0', '0-0-0-0']);
+  //   });
+
+  //   it('fires expand event', () => {
+  //     const handleExpand = jest.fn();
+  //     const wrapper = mount(
+  //       <Tree onExpand={handleExpand}>
+  //         <TreeNode title="parent 1" key="0-0">
+  //           <TreeNode title="leaf 1" key="0-0-0" />
+  //         </TreeNode>
+  //       </Tree>,
+  //     );
+  //     const switcher = wrapper.find('.rc-tree-switcher');
+
+  //     let node = convertNodePropsToEventData(wrapper.find(InternalTreeNode).first().props());
+  //     switcher.simulate('click');
+  //     expect(handleExpand).toHaveBeenCalledWith(['0-0'], {
+  //       expanded: true,
+  //       node,
+  //       nativeEvent: expect.anything(),
+  //     });
+
+  //     node = convertNodePropsToEventData(wrapper.find(InternalTreeNode).first().props());
+  //     switcher.simulate('click');
+  //     expect(handleExpand).toHaveBeenCalledWith([], {
+  //       expanded: false,
+  //       node,
+  //       nativeEvent: expect.anything(),
+  //     });
+  //   });
+  // });
+
+  describe('check', () => {
+    it('basic render', () => {
+      const { asFragment } = render(
+        <Tree checkable defaultExpandAll>
+          <TreeNode key="0-0">
+            <TreeNode key="0-0-0" disabled />
+            <TreeNode key="0-0-1" />
+          </TreeNode>
+        </Tree>,
+      );
+
+      expect(asFragment().firstChild).toMatchSnapshot();
+    });
+
+    //   it('checks default checked keys', () => {
+    //     const wrapper = mount(
+    //       <Tree checkable defaultCheckedKeys={['0-0']}>
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="leaf 1" key="0-0-0" />
+    //         </TreeNode>
+    //       </Tree>,
+    //     );
+    //     wrapper.find('.rc-tree-switcher').simulate('click');
+    //     wrapper.find('.rc-tree-checkbox').forEach(checkbox => {
+    //       expect(checkbox.is(CHECKED_CLASSNAME)).toBe(true);
+    //     });
+    //   });
+
+    //   it("ignore disabled children when calculate parent's checked status", () => {
+    //     const wrapper = mount(
+    //       <Tree checkable defaultCheckedKeys={['0-0-0']}>
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="leaf 1" key="0-0-0" disableCheckbox />
+    //           <TreeNode title="leaf 1" key="0-0-1" />
+    //         </TreeNode>
+    //       </Tree>,
+    //     );
+    //     const firstCheckboxClassNames = wrapper.find('.rc-tree-checkbox').instance().classList;
+    //     expect([].slice.call(firstCheckboxClassNames).includes(CHECKED_CLASSNAME.slice(1))).toBe(
+    //       false,
+    //     );
+    //   });
+
+    //   it('controlled by checkedKeys', () => {
+    //     const wrapper = mount(
+    //       <Tree checkable checkedKeys={[]}>
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="leaf 1" key="0-0-0" />
+    //         </TreeNode>
+    //       </Tree>,
+    //     );
+    //     const getCheckbox = () => wrapper.find('.rc-tree-checkbox');
+    //     expect(getCheckbox().is(CHECKED_CLASSNAME)).toBe(false);
+    //     wrapper.setProps({ checkedKeys: ['0-0'] });
+    //     expect(getCheckbox().is(CHECKED_CLASSNAME)).toBe(true);
+    //   });
+
+    //   it('trurns parent node to checked when all children are checked', () => {
+    //     const wrapper = mount(
+    //       <Tree checkable>
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="leaf 1" key="0-0-0" />
+    //           <TreeNode title="leaf 2" key="0-0-1" />
+    //         </TreeNode>
+    //       </Tree>,
+    //     );
+    //     wrapper.find('.rc-tree-switcher').simulate('click');
+    //     wrapper.find('.rc-tree-checkbox').at(1).simulate('click');
+    //     wrapper.find('.rc-tree-checkbox').at(2).simulate('click');
+    //     expect(wrapper.find('.rc-tree-checkbox').first().is(CHECKED_CLASSNAME)).toBe(true);
+    //   });
+
+    //   it('turns parent node to half checked when child is checked', () => {
+    //     const wrapper = mount(
+    //       <Tree checkable>
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="leaf 1" key="0-0-0" />
+    //           <TreeNode title="leaf 2" key="0-0-1" />
+    //         </TreeNode>
+    //       </Tree>,
+    //     );
+    //     wrapper.find('.rc-tree-switcher').simulate('click');
+    //     wrapper.find('.rc-tree-checkbox').last().simulate('click');
+    //     expect(wrapper.find('.rc-tree-checkbox').first().is('.rc-tree-checkbox-indeterminate')).toBe(
+    //       true,
+    //     );
+    //   });
+
+    //   it('turns parent node to checked if it is half checked', () => {
+    //     const wrapper = mount(
+    //       <Tree checkable>
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="leaf 1" key="0-0-0" />
+    //           <TreeNode title="leaf 2" key="0-0-1" />
+    //         </TreeNode>
+    //       </Tree>,
+    //     );
+    //     wrapper.find('.rc-tree-switcher').simulate('click');
+    //     wrapper.find('.rc-tree-checkbox').last().simulate('click');
+    //     wrapper.find('.rc-tree-checkbox').first().simulate('click');
+    //     wrapper.find('.rc-tree-checkbox').forEach(checkbox => {
+    //       expect(checkbox.is(CHECKED_CLASSNAME)).toBe(true);
+    //     });
+    //   });
+
+    //   it('fires check event', () => {
+    //     const handleCheck = jest.fn();
+    //     const wrapper = mount(
+    //       <Tree checkable onCheck={handleCheck}>
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="leaf 1" key="0-0-0" />
+    //         </TreeNode>
+    //       </Tree>,
+    //     );
+    //     wrapper.find('.rc-tree-switcher').simulate('click');
+
+    //     const nodeData00 = { title: 'parent 1', key: '0-0' };
+    //     const nodeData000 = { title: 'leaf 1', key: '0-0-0' };
+
+    //     let node = convertNodePropsToEventData(wrapper.find(InternalTreeNode).first().props());
+    //     wrapper.find('.rc-tree-checkbox').first().simulate('click');
+    //     expect(handleCheck).toHaveBeenCalledWith(
+    //       ['0-0', '0-0-0'],
+    //       objectMatcher({
+    //         checked: true,
+    //         checkedNodes: [nodeData00, nodeData000],
+    //         checkedNodesPositions: [
+    //           { node: nodeData00, pos: '0-0' },
+    //           { node: nodeData000, pos: '0-0-0' },
+    //         ],
+    //         event: 'check',
+    //         halfCheckedKeys: [],
+    //         node,
+    //         nativeEvent: {},
+    //       }),
+    //     );
+
+    //     node = convertNodePropsToEventData(wrapper.find(InternalTreeNode).last().props());
+    //     wrapper.find('.rc-tree-checkbox').last().simulate('click');
+    //     expect(handleCheck).toHaveBeenCalledWith(
+    //       [],
+    //       objectMatcher({
+    //         checked: false,
+    //         checkedNodes: [],
+    //         checkedNodesPositions: [],
+    //         event: 'check',
+    //         halfCheckedKeys: [],
+    //         node,
+    //         nativeEvent: {},
+    //       }),
+    //     );
+    //   });
+
+    //   // https://github.com/react-component/tree/issues/117
+    //   it('check works correctly after dragging children under another node', () => {
+    //     const wrapper = mount(
+    //       <Tree defaultExpandAll checkable>
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="leaf 1" key="0-0-0" />
+    //           <TreeNode title="leaf 2" key="0-0-1" />
+    //         </TreeNode>
+    //       </Tree>,
+    //     );
+    //     wrapper.find('.rc-tree-checkbox').at(1).simulate('click');
+    //     wrapper.setProps({
+    //       children: (
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="leaf 2" key="0-0-1">
+    //             <TreeNode title="leaf 1" key="0-0-0" />
+    //           </TreeNode>
+    //         </TreeNode>
+    //       ),
+    //     });
+    //     expect(() => wrapper.find('.rc-tree-checkbox').at(2).simulate('click')).not.toThrow();
+    //   });
+    //   // https://github.com/react-component/tree/issues/90
+    //   it('check works correctly after adding children dynamically', () => {
+    //     const wrapper = mount(
+    //       <Tree defaultExpandAll checkable>
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="leaf 1" key="0-0-0" />
+    //         </TreeNode>
+    //       </Tree>,
+    //     );
+    //     wrapper.find('.rc-tree-checkbox').at(1).simulate('click');
+    //     wrapper.setProps({
+    //       children: (
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="leaf 1" key="0-0-0" />
+    //           <TreeNode title="leaf 2" key="0-0-1" />
+    //         </TreeNode>
+    //       ),
+    //     });
+    //     expect(() => wrapper.find('.rc-tree-checkbox').at(2).simulate('click')).not.toThrow();
+    //   });
+
+    //   // https://github.com/ant-design/ant-design/issues/7353
+    //   it('check children after changing from children[disableCheckbox] from true to false', () => {
+    //     let checkedKeys = null;
+    //     const mockHandleCheck = keys => (checkedKeys = keys);
+    //     function Test({ disableCheckbox }) {
+    //       return (
+    //         <Tree checkable onCheck={mockHandleCheck}>
+    //           <TreeNode title="parent 1" key="0-0">
+    //             <TreeNode title="leaf 1" key="0-0-1" disableCheckbox={disableCheckbox} />
+    //             <TreeNode title="leaf 2" key="0-0-2" disableCheckbox={disableCheckbox} />
+    //             <TreeNode title="leaf 3" key="0-0-3" disableCheckbox={disableCheckbox} />
+    //           </TreeNode>
+    //         </Tree>
+    //       );
+    //     }
+    //     const wrapper = mount(<Test disableCheckbox />);
+    //     wrapper.find('.rc-tree-checkbox').first().simulate('click');
+    //     expect(checkedKeys).toEqual(['0-0']);
+    //     wrapper.setProps({ disableCheckbox: false });
+    //     wrapper.find('.rc-tree-checkbox').first().simulate('click').simulate('click');
+    //     expect(checkedKeys).toEqual(['0-0', '0-0-1', '0-0-2', '0-0-3']);
+    //   });
+
+    //   it('check dynamic children when their parent is checked', () => {
+    //     function mockLoadData() {}
+    //     const wrapper = mount(
+    //       <Tree checkable defaultCheckedKeys={['0-0']} loadData={mockLoadData}>
+    //         <TreeNode title="parent 1" key="0-0" />
+    //       </Tree>,
+    //     );
+    //     wrapper.setProps({
+    //       children: (
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="leaf 1" key="0-0-0" />
+    //           <TreeNode title="leaf 2" key="0-0-1" />
+    //         </TreeNode>
+    //       ),
+    //     });
+    //     expect(wrapper.state('checkedKeys')).toEqual(['0-0', '0-0-0', '0-0-1']);
+    //   });
+
+    // https://github.com/ant-design/ant-design/issues/10132
+    it('check update when Tree trigger componentWillReceiveProps', () => {
+      class Test extends React.Component {
+        state = {};
+
+        onCheck = () => {
+          this.setState({ whatever: 1 });
+        };
+
+        render() {
+          return (
+            <Tree checkable>
+              <TreeNode title="parent 1" key="0-0" />
+            </Tree>
+          );
+        }
+      }
+      const { container } = render(<Test />);
+      // wrapper.find('.rc-tree-checkbox').first().simulate('click');
+      fireEvent.click(container.querySelector('.rc-tree-checkbox'));
+
+      expect(container.firstChild).toMatchSnapshot();
+    });
+
+    describe('check after data ready', () => {
+      const errorSpy = spyError();
+
+      it('works', () => {
+        const checkedKeys = ['0-0-0'];
+        const { container, rerender } = render(<Tree checkable checkedKeys={checkedKeys} />);
+        expect(errorSpy()).toHaveBeenCalledWith("Warning: Tree missing follow keys: '0-0-0'");
+
+        rerender(
+          <Tree checkable checkedKeys={checkedKeys} expandedKeys={['0-0']}>
+            <TreeNode key="0-0" title="Light">
+              <TreeNode key="0-0-0" title="Bamboo" />
+            </TreeNode>
+          </Tree>,
+        );
+
+        expect(container.firstChild).toMatchSnapshot();
+      });
+    });
+
+    //   it('should ignore !checkable node', () => {
+    //     const onCheck = jest.fn();
+    //     const wrapper = mount(
+    //       <Tree checkable defaultExpandAll onCheck={onCheck}>
+    //         <TreeNode key="0-0">
+    //           <TreeNode key="0-0-0" checkable={false} />
+    //           <TreeNode key="0-0-1" />
+    //         </TreeNode>
+    //       </Tree>,
+    //     );
+
+    //     wrapper.find('.rc-tree-checkbox').last().simulate('click');
+
+    //     expect(onCheck.mock.calls[0][0].sort()).toEqual(['0-0', '0-0-1']);
+    //   });
+
+    //   describe('strictly', () => {
+    //     it('checks strictly', () => {
+    //       const wrapper = mount(
+    //         <Tree checkable checkStrictly>
+    //           <TreeNode title="parent 1" key="0-0">
+    //             <TreeNode title="leaf 1" key="0-0-0" />
+    //           </TreeNode>
+    //         </Tree>,
+    //       );
+    //       wrapper.find('.rc-tree-switcher').simulate('click');
+    //       wrapper.find('.rc-tree-checkbox').first().simulate('click');
+    //       expect(wrapper.find('.rc-tree-checkbox').first().is(CHECKED_CLASSNAME)).toBe(true);
+    //       expect(wrapper.find('.rc-tree-checkbox').last().is(CHECKED_CLASSNAME)).toBe(false);
+    //     });
+
+    //     describe('controlled mode', () => {
+    //       class App extends React.Component {
+    //         state = {
+    //           checkedKeys: {
+    //             checked: [],
+    //             halfChecked: [],
+    //           },
+    //         };
+
+    //         handleCheck = checkedKeys => {
+    //           this.setState({ checkedKeys });
+    //         };
+
+    //         render() {
+    //           return (
+    //             <Tree
+    //               checkable
+    //               checkStrictly
+    //               checkedKeys={this.state.checkedKeys}
+    //               onCheck={this.handleCheck}
+    //             >
+    //               <TreeNode title="parent 1" key="0-0">
+    //                 <TreeNode title="leaf 1" key="0-0-0" />
+    //               </TreeNode>
+    //             </Tree>
+    //           );
+    //         }
+    //       }
+
+    //       it('do not check children', () => {
+    //         const wrapper = mount(<App />);
+    //         wrapper.find('.rc-tree-switcher').simulate('click');
+    //         wrapper.find('.rc-tree-checkbox').first().simulate('click');
+    //         expect(wrapper.find('.rc-tree-checkbox').first().is(CHECKED_CLASSNAME)).toBe(true);
+    //         expect(wrapper.find('.rc-tree-checkbox').last().is(CHECKED_CLASSNAME)).toBe(false);
+    //       });
+
+    //       it('do not uncheck parent', () => {
+    //         const wrapper = mount(<App />);
+    //         wrapper.setState({
+    //           checkedKeys: { checked: ['0-0', '0-0-0'], halfChecked: [] },
+    //         });
+    //         wrapper.find('.rc-tree-switcher').simulate('click');
+    //         wrapper.find('.rc-tree-checkbox').last().simulate('click');
+    //         expect(wrapper.find('.rc-tree-checkbox').first().is(CHECKED_CLASSNAME)).toBe(true);
+    //         expect(wrapper.find('.rc-tree-checkbox').last().is(CHECKED_CLASSNAME)).toBe(false);
+    //       });
+    //     });
+    //   });
+
+    //   // https://github.com/ant-design/ant-design/issues/6891
+    //   it('should ignore disableCheckbox children items when check parent', () => {
+    //     const wrapper = mount(
+    //       <Tree checkable defaultExpandAll>
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="node" key="0-0-0" disableCheckbox>
+    //             <TreeNode title="node" key="0-0-0-0" disableCheckbox />
+    //             <TreeNode title="node" key="0-0-0-1" disableCheckbox />
+    //           </TreeNode>
+    //           <TreeNode title="node" key="0-0-1" />
+    //           <TreeNode title="node" key="0-0-2" />
+    //         </TreeNode>
+    //       </Tree>,
+    //     );
+
+    //     const firstNode = () => wrapper.find('.rc-tree-checkbox').first();
+    //     firstNode().simulate('click');
+
+    //     expect(firstNode().hasClass('rc-tree-checkbox-checked')).toBe(true);
+    //     expect(firstNode().hasClass('rc-tree-checkbox-indeterminate')).toBe(false);
+    //     expect(wrapper.state().checkedKeys.slice().sort()).toEqual(['0-0-1', '0-0-2', '0-0'].sort());
+
+    //     firstNode().simulate('click');
+    //     expect(firstNode().hasClass('rc-tree-checkbox-checked')).toBe(false);
+    //     expect(firstNode().hasClass('rc-tree-checkbox-indeterminate')).toBe(false);
+    //     expect(wrapper.state().checkedKeys).toEqual([]);
+    //   });
+
+    //   it('should ignore disabled children items when check parent', () => {
+    //     const wrapper = mount(
+    //       <Tree checkable defaultExpandAll>
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="node" key="0-0-0" disabled>
+    //             <TreeNode title="node" key="0-0-0-0" disabled />
+    //             <TreeNode title="node" key="0-0-0-1" disabled />
+    //           </TreeNode>
+    //           <TreeNode title="node" key="0-0-1" />
+    //           <TreeNode title="node" key="0-0-2" />
+    //         </TreeNode>
+    //       </Tree>,
+    //     );
+
+    //     const firstNode = () => wrapper.find(TreeNode).first().find('.rc-tree-checkbox').first();
+
+    //     firstNode().simulate('click');
+    //     expect(firstNode().hasClass('rc-tree-checkbox-checked')).toBe(true);
+    //     expect(firstNode().hasClass('rc-tree-checkbox-indeterminate')).toBe(false);
+    //     expect(wrapper.state().checkedKeys.slice().sort()).toEqual(['0-0-1', '0-0-2', '0-0'].sort());
+
+    //     firstNode().simulate('click');
+    //     expect(firstNode().hasClass('rc-tree-checkbox-checked')).toBe(false);
+    //     expect(firstNode().hasClass('rc-tree-checkbox-indeterminate')).toBe(false);
+    //     expect(wrapper.state().checkedKeys).toEqual([]);
+    //   });
+
+    //   // https://github.com/react-component/tree/pull/106#issuecomment-316779889
+    //   it('should render parent check state correctly', () => {
+    //     const wrapper = mount(
+    //       <Tree checkable defaultExpandAll>
+    //         <TreeNode title="parent 1" key="0-0">
+    //           <TreeNode title="node" key="0-0-0">
+    //             <TreeNode title="node" key="0-0-0-0" />
+    //             <TreeNode title="node" key="0-0-0-1" disableCheckbox />
+    //           </TreeNode>
+    //           <TreeNode title="node" key="0-0-1" />
+    //           <TreeNode title="node" key="0-0-2" />
+    //         </TreeNode>
+    //       </Tree>,
+    //     );
+    //     const getParent = () => wrapper.find(TreeNode).at(1).find('.rc-tree-checkbox').first();
+    //     const getNode = () => wrapper.find(TreeNode).at(2).find('.rc-tree-checkbox').first();
+
+    //     getNode().simulate('click');
+    //     expect(getNode().hasClass('rc-tree-checkbox-checked')).toBe(true);
+    //     expect(getParent().hasClass('rc-tree-checkbox-indeterminate')).toBe(false);
+    //     expect(getParent().hasClass('rc-tree-checkbox-checked')).toBe(true);
+    //     getNode().simulate('click');
+    //     expect(getNode().hasClass('rc-tree-checkbox-checked')).toBe(false);
+    //     expect(getParent().hasClass('rc-tree-checkbox-indeterminate')).toBe(false);
+    //     expect(getParent().hasClass('rc-tree-checkbox-checked')).toBe(false);
+    //   });
+  });
+
+  // describe('select', () => {
+  //   it('selects default selected keys', () => {
+  //     const wrapper = mount(
+  //       <Tree selectable defaultSelectedKeys={['0-0']}>
+  //         <TreeNode title="parent 1" key="0-0">
+  //           <TreeNode title="leaf 1" key="0-0-0" />
+  //         </TreeNode>
+  //       </Tree>,
+  //     );
+  //     expect(wrapper.find('.rc-tree-node-content-wrapper').is(SELECTED_CLASSNAME)).toBe(true);
+  //   });
+
+  //   it('controlled by selectedKeys', () => {
+  //     const wrapper = mount(
+  //       <Tree selectable selectedKeys={[]}>
+  //         <TreeNode title="parent 1" key="0-0">
+  //           <TreeNode title="leaf 1" key="0-0-0" />
+  //         </TreeNode>
+  //       </Tree>,
+  //     );
+  //     expect(wrapper.find('.rc-tree-node-content-wrapper').is(SELECTED_CLASSNAME)).toBe(false);
+  //     wrapper.setProps({ selectedKeys: ['0-0'] });
+  //     expect(wrapper.find('.rc-tree-node-content-wrapper').is(SELECTED_CLASSNAME)).toBe(true);
+  //   });
+
+  //   it('fires select event', () => {
+  //     const handleSelect = jest.fn();
+  //     const wrapper = mount(
+  //       <Tree selectable onSelect={handleSelect}>
+  //         <TreeNode title="parent 1" key="0-0">
+  //           <TreeNode title="leaf 1" key="0-0-0" />
+  //         </TreeNode>
+  //       </Tree>,
+  //     );
+  //     const nodeContent = wrapper.find('.rc-tree-node-content-wrapper');
+
+  //     let node = convertNodePropsToEventData(wrapper.find(InternalTreeNode).first().props());
+  //     nodeContent.simulate('click');
+  //     expect(handleSelect).toHaveBeenCalledWith(
+  //       ['0-0'],
+  //       objectMatcher({
+  //         event: 'select',
+  //         node,
+  //         selected: true,
+  //         selectedNodes: [{ title: 'parent 1', key: '0-0' }],
+  //         nativeEvent: {},
+  //       }),
+  //     );
+
+  //     node = convertNodePropsToEventData(wrapper.find(InternalTreeNode).first().props());
+  //     nodeContent.simulate('click');
+  //     expect(handleSelect).toHaveBeenCalledWith(
+  //       [],
+  //       objectMatcher({
+  //         event: 'select',
+  //         node,
+  //         selected: false,
+  //         selectedNodes: [],
+  //         nativeEvent: {},
+  //       }),
+  //     );
+  //   });
+  // });
+
+  // describe('checkable but not selectable', () => {
+  //   it('fires check event when click on TreeNode', () => {
+  //     const handleCheck = jest.fn();
+  //     const wrapper = mount(
+  //       <Tree checkable selectable={false} onCheck={handleCheck}>
+  //         <TreeNode title="parent 1" key="0-0">
+  //           <TreeNode title="leaf 1" key="0-0-0" />
+  //         </TreeNode>
+  //       </Tree>,
+  //     );
+  //     wrapper.find('.rc-tree-switcher').simulate('click');
+
+  //     const dataNode1 = { title: 'parent 1', key: '0-0' };
+  //     const dataNode2 = { title: 'leaf 1', key: '0-0-0' };
+
+  //     let node = convertNodePropsToEventData(wrapper.find(InternalTreeNode).first().props());
+  //     wrapper.find('.rc-tree-node-content-wrapper').first().simulate('click');
+  //     expect(handleCheck).toHaveBeenCalledWith(
+  //       ['0-0', '0-0-0'],
+  //       objectMatcher({
+  //         checked: true,
+  //         checkedNodes: [dataNode1, dataNode2],
+  //         checkedNodesPositions: [
+  //           { node: dataNode1, pos: '0-0' },
+  //           { node: dataNode2, pos: '0-0-0' },
+  //         ],
+  //         event: 'check',
+  //         halfCheckedKeys: [],
+  //         node,
+  //         nativeEvent: {},
+  //       }),
+  //     );
+
+  //     node = convertNodePropsToEventData(wrapper.find(InternalTreeNode).last().props());
+  //     wrapper.find('.rc-tree-node-content-wrapper').last().simulate('click');
+  //     expect(handleCheck).toHaveBeenCalledWith(
+  //       [],
+  //       objectMatcher({
+  //         checked: false,
+  //         checkedNodes: [],
+  //         checkedNodesPositions: [],
+  //         event: 'check',
+  //         halfCheckedKeys: [],
+  //         node,
+  //         nativeEvent: {},
+  //       }),
+  //     );
+  //   });
+  // });
+
+  // it('fires rightClick event', () => {
+  //   const handleRightClick = jest.fn();
+  //   const wrapper = mount(
+  //     <Tree onRightClick={handleRightClick}>
+  //       <TreeNode title="parent 1" key="0-0">
+  //         <TreeNode title="leaf 1" key="0-0-0" />
+  //       </TreeNode>
+  //     </Tree>,
+  //   );
+  //   const node = convertNodePropsToEventData(wrapper.find(InternalTreeNode).props());
+  //   wrapper.find('.rc-tree-node-content-wrapper').simulate('contextMenu');
+  //   expect(handleRightClick.mock.calls[0][0].node).toEqual(node);
+  // });
+
+  // it('fires rightClick should not change selected item', () => {
+  //   const handleRightClick = jest.fn();
+  //   const wrapper = mount(
+  //     <Tree onRightClick={handleRightClick}>
+  //       <TreeNode title="parent 1" key="0-0">
+  //         <TreeNode title="leaf 1" key="0-0-0" />
+  //       </TreeNode>
+  //     </Tree>,
+  //   );
+  //   wrapper.find('.rc-tree-node-content-wrapper').simulate('contextMenu');
+  //   expect(wrapper.state().selectedKeys.length).toBe(0);
+  // });
+
+  // it('fires mouseEnter events', () => {
+  //   const handleMouseEnter = jest.fn();
+  //   const wrapper = mount(
+  //     <Tree onMouseEnter={handleMouseEnter}>
+  //       <TreeNode title="parent 1" key="0-0">
+  //         <TreeNode title="leaf 1" key="0-0-0" />
+  //       </TreeNode>
+  //     </Tree>,
+  //   );
+  //   const node = convertNodePropsToEventData(wrapper.find(InternalTreeNode).props());
+  //   wrapper.find('.rc-tree-node-content-wrapper').simulate('mouseEnter');
+  //   expect(handleMouseEnter.mock.calls[0][0].node).toEqual(node);
+  // });
+
+  // it('fires mouseLeave events', () => {
+  //   const handleMouseLeave = jest.fn();
+  //   const wrapper = mount(
+  //     <Tree onMouseLeave={handleMouseLeave}>
+  //       <TreeNode title="parent 1" key="0-0">
+  //         <TreeNode title="leaf 1" key="0-0-0" />
+  //       </TreeNode>
+  //     </Tree>,
+  //   );
+  //   const node = convertNodePropsToEventData(wrapper.find(InternalTreeNode).props());
+  //   wrapper.find('.rc-tree-node-content-wrapper').simulate('mouseLeave');
+  //   expect(handleMouseLeave.mock.calls[0][0].node).toEqual(node);
+  // });
+
+  // it('filters nodes', () => {
+  //   function filterTreeNode(treeNode) {
+  //     return treeNode.title.indexOf('parent') > -1;
+  //   }
+  //   const wrapper = mount(
+  //     <Tree filterTreeNode={filterTreeNode}>
+  //       <TreeNode title="parent 1" key="0-0">
+  //         <TreeNode title="leaf 1" key="0-0-0">
+  //           <TreeNode title="meck" key="0-0-0-0" />
+  //           <TreeNode title="leaf" key="0-0-0-1" />
+  //         </TreeNode>
+  //         <TreeNode title="leaf 2" key="0-0-1" />
+  //       </TreeNode>
+  //     </Tree>,
+  //   );
+
+  //   expect(wrapper.find(InternalTreeNode).find('div').is('.filter-node')).toBe(true);
+  // });
+
+  // it('loads nodes asynchronously', () => {
+  //   const then = jest.fn(() => Promise.resolve());
+  //   const loadData = jest.fn(() => ({ then }));
+  //   const wrapper = mount(
+  //     <Tree loadData={loadData}>
+  //       <TreeNode title="parent 1" key="0-0" />
+  //     </Tree>,
+  //   );
+  //   const node = convertNodePropsToEventData(wrapper.find(InternalTreeNode).props());
+  //   wrapper.find('.rc-tree-switcher').simulate('click');
+  //   expect(loadData).toHaveBeenCalledWith(node);
+  //   expect(then).toHaveBeenCalled();
+  // });
+
+  // it('renders without errors', () => {
+  //   expect(() => {
+  //     mount(
+  //       <Tree>
+  //         {[0, 1].map(i => (
+  //           <TreeNode title={i} key={i}>
+  //             {[2, 3].map(j => (
+  //               <TreeNode title={j} key={`${i}_${j}`} />
+  //             ))}
+  //             <TreeNode title="4" key={`${i}_4`} />
+  //           </TreeNode>
+  //         ))}
+  //       </Tree>,
+  //     );
+  //   }).not.toThrow();
+  // });
+
+  it('renders opaque children correctly', () => {
+    const { asFragment } = render(
+      <Tree>
+        <TreeNode title="0" key="0">
+          {[1, 2].map(i => (
+            <TreeNode title={i} key={i} />
+          ))}
+          <TreeNode title="3" key="3" />
+        </TreeNode>
+      </Tree>,
+    );
+
+    expect(asFragment().firstChild).toMatchSnapshot();
+  });
+
+  describe('ignore illegal node as Tree children', () => {
+    const errorSpy = spyError();
+
+    it('Direct TreeNode', () => {
+      const { asFragment } = render(
+        <Tree defaultExpandAll>
+          <TreeNode key="00" title="00" />
+          <span>Hide Me</span>
+          <TreeNode key="02" title="02" />
+        </Tree>,
+      );
+      expect(asFragment().firstChild).toMatchSnapshot();
+      expect(errorSpy()).toHaveBeenCalledWith(
+        'Warning: Tree/TreeNode can only accept TreeNode as children.',
+      );
+    });
+
+    it('Sub TreeNode', () => {
+      resetWarned();
+      const { asFragment } = render(
+        <Tree defaultExpandAll>
+          <TreeNode key="00" title="00" />
+          <TreeNode key="01" title="01">
+            <TreeNode key="010" title="010" />
+            <span>I AM INVISIBLE</span>
+            <TreeNode key="012" title="012" />
+          </TreeNode>
+        </Tree>,
+      );
+      expect(asFragment().firstChild).toMatchSnapshot();
+      expect(errorSpy()).toHaveBeenCalledWith(
+        'Warning: Tree/TreeNode can only accept TreeNode as children.',
+      );
+    });
+  });
+
+  // it('Number key', () => {
+  //   const onCheck = jest.fn();
+
+  //   const Demo = () => {
+  //     const [checkedKeys, setCheckedKeys] = React.useState([]);
+
+  //     return (
+  //       <Tree
+  //         checkable
+  //         defaultExpandAll
+  //         checkedKeys={checkedKeys}
+  //         onCheck={keys => {
+  //           setCheckedKeys(keys);
+  //           onCheck(keys);
+  //         }}
+  //         treeData={[{ key: 11, title: 11, children: [{ key: 22, title: 22 }] }]}
+  //       />
+  //     );
+  //   };
+
+  //   const wrapper = mount(<Demo />);
+  //   wrapper.update();
+
+  //   wrapper.find('.rc-tree-checkbox').first().simulate('click');
+  //   expect(onCheck).toHaveBeenCalledWith([11, 22]);
+
+  //   onCheck.mockReset();
+  //   wrapper.find('.rc-tree-checkbox').last().simulate('click');
+  //   expect(onCheck).toHaveBeenCalledWith([]);
+  // });
+
+  // describe('scrollTo should work', () => {
+  //   let domSpy;
+  //   let called = false;
+
+  //   beforeAll(() => {
+  //     domSpy = spyElementPrototypes(HTMLDivElement, {
+  //       scrollTop: {
+  //         get: () => 233,
+  //         set: () => {
+  //           called = true;
+  //         },
+  //       },
+  //       clientHeight: {
+  //         get: () => 100,
+  //       },
+  //     });
+  //   });
+
+  //   afterAll(() => {
+  //     domSpy.mockRestore();
+  //   });
+
+  //   it('work', () => {
+  //     jest.useFakeTimers();
+  //     const wrapper = mount(<Tree />);
+
+  //     wrapper.instance().scrollTo({ key: 'light', align: 'top' });
+  //     jest.runAllTimers();
+
+  //     expect(called).toBeTruthy();
+  //     jest.useRealTimers();
+  //   });
+  // });
+
+  // it('not crash if expandedKeys is null', () => {
+  //   mount(
+  //     <Tree expandedKeys={null}>
+  //       <TreeNode key="test" title="alive" />
+  //     </Tree>,
+  //   );
+  // });
+
+  // it('not crash if expandedKeys change to undefined', () => {
+  //   const wrapper = mount(
+  //     <Tree expandedKeys={['light']}>
+  //       <TreeNode key="light" title="light">
+  //         <TreeNode key="bamboo" title="bamboo" />
+  //       </TreeNode>
+  //     </Tree>,
+  //   );
+  //   const { data } = wrapper.find(VirtualList).props();
+  //   expect(data).toHaveLength(2);
+
+  //   wrapper.setProps({ expandedKeys: undefined });
+
+  //   // Click should not crash
+  //   wrapper.find('.rc-tree-switcher').first().simulate('click');
+  // });
+
+  // it('controlled should block expanded', () => {
+  //   const wrapper = mount(
+  //     <Tree
+  //       expandedKeys={['1']}
+  //       treeData={[{ key: '1', title: 1, children: [{ key: '2', title: 2 }] }]}
+  //     />,
+  //   );
+
+  //   wrapper.find('.rc-tree-switcher').first().simulate('click');
+  //   expect(wrapper.find('List').props().data).toHaveLength(2);
+  // });
+
+  // it('support virtual', () => {
+  //   const data = [];
+  //   for (let i = 0; i < 99; i += 1) {
+  //     data.push({
+  //       key: i,
+  //       title: i,
+  //     });
+  //   }
+
+  //   const wrapper = mount(<Tree itemHeight={10} height={100} treeData={data} virtual={false} />);
+
+  //   expect(wrapper.find('List').props().virtual).toBe(false);
+  // });
+
+  // // https://github.com/ant-design/ant-design/issues/28349
+  // it('should not trigger expend when loading data', () => {
+  //   const then = jest.fn(() => Promise.resolve());
+  //   const loadData = jest.fn(() => ({ then }));
+  //   const onExpand = jest.fn();
+  //   const wrapper = mount(
+  //     <Tree loadData={loadData} onExpand={onExpand}>
+  //       <TreeNode title="parent 1" key="0-0" />
+  //     </Tree>,
+  //   );
+
+  //   // trigger click to expand node
+  //   wrapper.find('.rc-tree-switcher').simulate('click');
+  //   expect(wrapper.find('.rc-tree-switcher').first().is(OPEN_CLASSNAME)).toBe(true);
+  //   expect(onExpand).toBeCalled();
+
+  //   // click again
+  //   onExpand.mockReset();
+  //   wrapper.find('.rc-tree-switcher').simulate('click');
+  //   expect(wrapper.find('.rc-tree-switcher').first().is(OPEN_CLASSNAME)).toBe(true);
+  //   expect(onExpand).not.toHaveBeenCalled();
+  // });
+
+  // // https://github.com/ant-design/ant-design/issues/31141
+  // it('should restore tree node loading and expand status when loadData fail', async () => {
+  //   const loadData = () => new Promise((_, reject) => setTimeout(reject));
+  //   const wrapper = mount(
+  //     <Tree loadData={loadData}>
+  //       <TreeNode title="parent 1" key="0-0" />
+  //     </Tree>,
+  //   );
+  //   // trigger click to expand node
+  //   wrapper.find('.rc-tree-switcher').simulate('click');
+  //   expect(
+  //     wrapper.find('.rc-tree-switcher').getDOMNode().className.includes('rc-tree-switcher_open'),
+  //   ).toBe(true);
+  //   await delay();
+  //   expect(
+  //     wrapper.find('.rc-tree-switcher').getDOMNode().className.includes('rc-tree-switcher_close'),
+  //   ).toBe(true);
+  // });
+
+  // // https://github.com/ant-design/ant-design/issues/31141
+  // it('should support onScroll', async () => {
+  //   const onScroll = jest.fn();
+  //   const data = [];
+  //   for (let i = 0; i < 99; i += 1) {
+  //     data.push({
+  //       key: i,
+  //       title: i,
+  //     });
+  //   }
+  //   const wrapper = mount(
+  //     <Tree itemHeight={10} height={100} treeData={data} onScroll={onScroll} />,
+  //   );
+  //   wrapper.find('.rc-tree-list-holder').simulate('scroll');
+  //   expect(onScroll).toBeCalled();
+  // });
+
+  it('should support rootStyle and rootClassName', () => {
+    const data = [];
+    for (let i = 0; i < 9; i += 1) {
+      data.push({
+        key: i,
+        title: i,
+      });
+    }
+    const { container } = render(
+      <Tree treeData={data} rootClassName="root-tree" rootStyle={{ backgroundColor: 'cyan' }} />,
+    );
+
+    expect(container.firstChild).toMatchSnapshot();
+    expect(container.firstChild).toHaveClass('root-tree');
+    expect(container.firstChild).toHaveStyle({ backgroundColor: 'cyan' });
+  });
+});
